requirements.txt
click==8.1.7
pypdf==4.3.1
python-dateutil==2.9.0.post0

